Feature/introduce db (#9) (#10)
* Added firestore to the waiting list component

* Made it so that the admin waiting list loads from firestore and modified
admin waiting list reducer to take care of the ad row functionality

* Started to move the add row stuff into redux so I can use it to add new
rows in firestore more easily. Not complete yet, still have to hookup
the action creators and address the linting errors

* Made it so that you can add scouts

* Fixed front end after column rename

* Changed fab so that we have a fab box instead which will act as a
toolbox for editing the waiting list

* Made the separation between authenticated admin waiting list and admin
waitin glist and also between the draggable row and the editable row.
Also added in a heap of tests

* Pulled the adding row out into it's own component and made it show up
when a new adding flag is set in state. This is all in prepration so
that I can make it so that you can either edit/delete or add but not
both to try and simplify what needs to be developed a bit.

* Started implementing deleting rows but it's got a bug might need to
delete the sensitive info first then delete the main to stop the issues

* Made deleting a row work

* Moved all the adding logic to the adding row

* Made it so that you can edit rows, did a quick bit of refactoring on the
add row naming and also make the with of the control column and date
format be driven by theme

* Made it so that the edit row changes are saved when hitting save

* In the middle of writing a test suite for the shuffling logic for
dragging rows. Currently the row dragging functionality just blats the
data. Need to finish the shuffling logic and then drop that in where the
drag end is

* Made the shuffling logic work but now there's a bug in redux firestore
where after you do an update the id property doesn't propagate itself
correctly

* Fixed final bugs in row switching, made it so that add row clears
itself, added missing fields to public list. A few other bugs as well

* Removing console logs and adding index